+++ src/main/kotlin/org/shangahi/sellio_backend/api/mapper/StoreCardResponseMapper.kt
@@ -2,12 +2,12 @@ package org.shangahi.sellio_backend.api.mapper
 
 import org.shangahi.sellio_backend.api.dto.request.StoreCardResponse
 import org.shangahi.sellio_backend.entity.Store
+import java.math.BigDecimal
 import java.util.UUID
-import kotlin.collections.contains
 import kotlin.collections.get
 
 fun Store.toStoreCardResponse(
-    discountsMap: Map<UUID, Double>,
+    discountsMap: Map<UUID, BigDecimal>,
     isFavorite: Boolean
 ): StoreCardResponse{
     return StoreCardResponse(
